getTextRecords should handle null input
