+++ src/api/manager/resolvers.js
@@ -38,41 +38,6 @@ async function delay(ms) {
   return await new Promise(resolve => setTimeout(resolve, ms))
 }
 
-function adjustForShortNames(node) {
-  const nameArray = node.name.split('.')
-  const { label, parent } = node
-
-  // return original node if is subdomain or not eth
-  if (nameArray.length > 2 || parent !== 'eth' || label.length > 6) return node
-
-  //if the auctions are over
-  if (new Date() > new Date(1570924800000)) {
-    return node
-  }
-
-  let auctionEnds
-  let onAuction = true
-
-  if (label.length >= 5) {
-    auctionEnds = new Date(1569715200000) // 29 September
-  } else if (label.length >= 4) {
-    auctionEnds = new Date(1570320000000) // 6 October
-  } else if (label.length >= 3) {
-    auctionEnds = new Date(1570924800000) // 13 October
-  }
-
-  if (new Date() > auctionEnds) {
-    onAuction = false
-  }
-
-  return {
-    ...node,
-    auctionEnds,
-    onAuction,
-    state: onAuction ? 'Auction' : node.state
-  }
-}
-
 function setState(node) {
   let state = node.state
   if (node.isDNSRegistrar) {
@@ -186,6 +151,185 @@ const handleMultipleTransactions = async (name, records, resolverInstance) => {
     return sendHelper(tx1)
   } catch (e) {
     console.log('error creating transaction array', e)
+  }
+}
+
+async function getRegistrarEntry(name) {
+  const registrar = getRegistrar()
+  const nameArray = name.split('.')
+  if (nameArray.length > 3 || nameArray[1] !== 'eth') {
+    return {}
+  }
+
+  const entry = await registrar.getEntry(nameArray[0])
+  const {
+    registrant,
+    deedOwner,
+    state,
+    registrationDate,
+    migrationStartDate,
+    currentBlockDate,
+    transferEndDate,
+    gracePeriodEndDate,
+    revealDate,
+    value,
+    highestBid,
+    expiryTime,
+    isNewRegistrar,
+    available
+  } = entry
+
+  const node = {
+    name: `${name}`,
+    state: modeNames[state],
+    stateError: null, // This is only used for dnssec errors
+    registrationDate,
+    gracePeriodEndDate: gracePeriodEndDate || null,
+    migrationStartDate: migrationStartDate || null,
+    currentBlockDate: currentBlockDate || null,
+    transferEndDate: transferEndDate || null,
+    revealDate,
+    value,
+    highestBid,
+    registrant,
+    deedOwner,
+    isNewRegistrar: !!isNewRegistrar,
+    available,
+    expiryTime: expiryTime || null
+  }
+
+  return node
+}
+
+async function getParent(name) {
+  const ens = getENS()
+  const nameArray = name.split('.')
+  if (nameArray.length < 1) {
+    return [null, null]
+  }
+  nameArray.shift()
+  const parent = nameArray.join('.')
+  const parentOwner = await ens.getOwner(parent)
+  return [parent, parentOwner]
+}
+
+async function getRegistrant(name) {
+  const client = getClient()
+  try {
+    const { data, error } = await client.query({
+      query: GET_REGISTRANT_FROM_SUBGRAPH,
+      fetchPolicy: 'network-only',
+      variables: { id: labelhash(name.split('.')[0]) }
+    })
+    if (!data || !data.registration) {
+      return null
+    }
+    if (error) {
+      console.log('Error getting registrant from subgraph', error)
+      return null
+    }
+
+    return utils.getAddress(data.registration.registrant.id)
+  } catch (e) {
+    console.log('GraphQL error from getRegistrant', e)
+    return null
+  }
+}
+
+async function getDNSEntryDetails(name) {
+  const ens = getENS()
+  const registrar = getRegistrar()
+  const nameArray = name.split('.')
+  const networkId = await getNetworkId()
+  if (nameArray.length !== 2 || nameArray[1] === 'eth') return {}
+
+  let tld = nameArray[1]
+  let owner
+  let tldowner
+  tldowner = (await ens.getOwner(tld)).toLocaleLowerCase()
+  if (parseInt(tldowner) === 0 && networkId === 3) {
+    tldowner = ROPSTEN_DNSREGISTRAR_ADDRESS
+  }
+
+  try {
+    owner = (await ens.getOwner(name)).toLocaleLowerCase()
+  } catch {
+    return {}
+  }
+
+  let isDNSRegistrarSupported = await registrar.isDNSRegistrar(tldowner)
+  if (isDNSRegistrarSupported && tldowner !== emptyAddress) {
+    const dnsEntry = await registrar.getDNSEntry(name, tldowner, owner)
+    const node = {
+      isDNSRegistrar: true,
+      dnsOwner: dnsEntry.claim?.result
+        ? dnsEntry.claim.getOwner()
+        : emptyAddress,
+      state: dnsEntry.state,
+      stateError: dnsEntry.stateError,
+      parentOwner: tldowner
+    }
+
+    return node
+  }
+}
+
+async function getTestEntry(name) {
+  const registrar = getRegistrar()
+  const nameArray = name.split('.')
+  if (nameArray.length < 3 && nameArray[1] === 'test') {
+    const expiryTime = await registrar.expiryTimes(nameArray[0])
+    if (expiryTime) return { expiryTime }
+  }
+  return {}
+}
+
+const waitForReactiveVar = varReactive =>
+  new Promise((resolve, reject) => {
+    let count = 0
+    const interval = setInterval(() => {
+      if (count === 5) {
+        reject('too many retries')
+      }
+      if (varReactive()) {
+        resolve(varReactive())
+      }
+      count++
+    }, 1000)
+  })
+
+function adjustForShortNames(node) {
+  const nameArray = node.name.split('.')
+  const { label, parent } = node
+
+  // return original node if is subdomain or not eth
+  if (nameArray.length > 2 || parent !== 'eth' || label.length > 6) return node
+
+  //if the auctions are over
+  if (new Date() > new Date(1570924800000)) {
+    return node
+  }
+
+  let auctionEnds
+  let onAuction = true
+
+  if (label.length >= 5) {
+    auctionEnds = new Date(1569715200000) // 29 September
+  } else if (label.length >= 4) {
+    auctionEnds = new Date(1570320000000) // 6 October
+  } else if (label.length >= 3) {
+    auctionEnds = new Date(1570924800000) // 13 October
+  }
+
+  if (new Date() > auctionEnds) {
+    onAuction = false
+  }
+
+  return {
+    ...node,
+    auctionEnds,
+    onAuction,
+    state: onAuction ? 'Auction' : node.state
   }
 }
 
@@ -479,6 +623,7 @@ const resolvers = {
       return test
     },
     getTextRecords: async (_, { name, keys }) => {
+      if (!name || !keys) return []
       const ens = getENS()
       const textRecords = keys.map(key =>
         ens.getText(name, key).then(addr => ({ key, value: addr }))
